requirements.txt
openai
ollama
tensorflow
numpy
editdistance
pandas
python-dotenv
google-genai